resolve Login nad SignUp rendering issue and forget-password
