+++ dashboard/static/assets/js/apps/CoreApp.js
@@ -6,6 +6,7 @@ import {
   ResourcesPageView,
   MutualPageView,
   ContactPageView,
+  ClientAuthView,
 } from "../components/core/App";
 import { Navbar } from "../components/core/coreComponents/navbar";
 
@@ -19,6 +20,12 @@ export function CoreApp() {
         <Route path="/resources" element={<ResourcesPageView />} />
         <Route path="/mutual-funds" element={<MutualPageView />} />
         <Route path="/contact" element={<ContactPageView />} />
+
+        <Route path="/register" element={<ClientAuthView formType="register"/>} />
+        <Route path="/login" element={<ClientAuthView formType="login"/>} />
+        <Route path="/password-reset" element={<ClientAuthView formType="password-reset"/>} />
+        
+
       </Routes>
     </>
   );
