+++ package.json
@@ -23,6 +23,7 @@
     "@react-icons/all-files": "^4.1.0",
     "@types/jest": "^29.5.2",
     "@typescript-eslint/parser": "^5.59.11",
+    "aos": "^2.3.4",
     "axios": "^1.4.0",
     "babel-cli": "^6.26.0",
     "babel-preset-react-app": "^3.1.2",
